fixed some bugs

+++ forum-api/src/main/java/com/platform/entity/TagTagVo.java
@@ -4,32 +4,36 @@ package com.platform.entity;
 import java.io.Serializable;
 
 public class TagTagVo implements Serializable {
-    private Integer oId;
-    private Integer tag1Oid;
-    private Integer tag2Oid;
+    //主键
+    private Long oId;
+    //标签1 id
+    private Long tag1Oid;
+    //标签2 id
+    private Long tag2Oid;
+    //权重
     private String weight;
 
-    public Integer getoId() {
+    public Long getoId() {
         return oId;
     }
 
-    public void setoId(Integer oId) {
+    public void setoId(Long oId) {
         this.oId = oId;
     }
 
-    public Integer getTag1Oid() {
+    public Long getTag1Oid() {
         return tag1Oid;
     }
 
-    public void setTag1Oid(Integer tag1Oid) {
+    public void setTag1Oid(Long tag1Oid) {
         this.tag1Oid = tag1Oid;
     }
 
-    public Integer getTag2Oid() {
+    public Long getTag2Oid() {
         return tag2Oid;
     }
 
-    public void setTag2Oid(Integer tag2Oid) {
+    public void setTag2Oid(Long tag2Oid) {
         this.tag2Oid = tag2Oid;
     }
